Add a try/catch to the controller call
We had found that when running Behat tests and we created a Page object as a fixture the onBeforeWrite would cascade to here. Given at the time of execution there is no controller it causes a failure in the Behat test.

This attempts to catch the exception of there being no current controller and handle it so that the Behat test can continue to run.